moves app layout into components/layout and other house cleaning
